Shrinking Circle for Stationary Tutorial Pointers

+++ Assets/Scripts/Database Scripts/DisplayOnceSystem.js
@@ -6,10 +6,13 @@ Created so that tutorial pointers and the inspection display only gets displayed
 
 */
 
+private var disableThisSystem: boolean = false;
+
 //Note, this will delete ALL PLAYERPREF KEYS
 public function DeleteKeys()
 {
-	PlayerPrefs.DeleteAll();
+	if(!disableThisSystem)
+		PlayerPrefs.DeleteAll();
 }
 
 
@@ -20,33 +23,47 @@ public function GetKeyString(index:int, isATutorialPointer:boolean):String
 
 public function WasAlreadyDisplayed(index:int, isATutorialPointer:boolean):boolean
 {
-	var name:String = GetKeyString(index, isATutorialPointer);
-	
-	if(PlayerPrefs.HasKey(name))
-		if(PlayerPrefs.GetInt(name) == 1)
-			return true;
-		else
-			return false;
+	if(!disableThisSystem)
+	{
+		var name:String = GetKeyString(index, isATutorialPointer);
+		
+		if(PlayerPrefs.HasKey(name))
+			if(PlayerPrefs.GetInt(name) == 1)
+				return true;
+			else
+				return false;
+	}
+	else
+		return false;
 }
 
 public function WasAlreadyDisplayed(name:String):boolean
 {	
-	if(PlayerPrefs.HasKey(name))
-		if(PlayerPrefs.GetInt(name) == 1)
-			return true;
-		else
-			return false;
+	if(!disableThisSystem)
+	{
+		if(PlayerPrefs.HasKey(name))
+			if(PlayerPrefs.GetInt(name) == 1)
+				return true;
+			else
+				return false;
+	}
+	else
+		return false;
 }
 
 public function HasDisplayed(name:String)
 {	
-	PlayerPrefs.SetInt(name, 1);
+	if(!disableThisSystem)
+		PlayerPrefs.SetInt(name, 1);
 }
 
 public function HasDisplayed(index:int, isATutorialPointer:boolean)
 {	
-	var name:String = GetKeyString(index, isATutorialPointer);
-	PlayerPrefs.SetInt(name,1);
+	if(!disableThisSystem)
+	{
+		var name:String = GetKeyString(index, isATutorialPointer);
+		PlayerPrefs.SetInt(name,1);
+	}
 }
